+++ backend/canisters/community/impl/src/lib.rs
@@ -7,7 +7,7 @@ use activity_notification_state::ActivityNotificationState;
 use candid::Principal;
 use canister_state_macros::canister_state;
 use canister_timer_jobs::TimerJobs;
-use chat_events::ChatMetricsInternal;
+use chat_events::{ChannelThreadKeyPrefix, ChatMetricsInternal, KeyPrefix};
 use community_canister::EventsResponse;
 use event_store_producer::{EventStoreClient, EventStoreClientBuilder, EventStoreClientInfo};
 use event_store_producer_cdk_runtime::CdkRuntime;
@@ -244,7 +244,16 @@ impl RuntimeState {
                 }
             }
             prize_refunds.extend(result.prize_refunds);
+            for (message_index, _) in result.threads {
+                self.data
+                    .stable_memory_keys_to_garbage_collect
+                    .push(KeyPrefix::ChannelThread(ChannelThreadKeyPrefix::new(
+                        channel.id,
+                        message_index,
+                    )));
+            }
         }
+        jobs::garbage_collect_stable_memory::start_job_if_required(self);
 
         self.data.next_event_expiry = next_event_expiry;
         if let Some(expiry) = self.data.next_event_expiry {
@@ -293,6 +302,7 @@ impl RuntimeState {
             instruction_counts: self.data.instruction_counts_log.iter().collect(),
             event_store_client_info: self.data.event_store_client.info(),
             timer_jobs: self.data.timer_jobs.len() as u32,
+            stable_memory_event_migration_complete: self.data.stable_memory_event_migration_complete,
             canister_ids: CanisterIds {
                 user_index: self.data.user_index_canister_id,
                 group_index: self.data.group_index_canister_id,
@@ -360,6 +370,10 @@ struct Data {
     expiring_member_actions: ExpiringMemberActions,
     user_cache: UserCache,
     user_event_sync_queue: GroupedTimerJobQueue<UserEventBatch>,
+    #[serde(default)]
+    stable_memory_event_migration_complete: bool,
+    #[serde(default)]
+    stable_memory_keys_to_garbage_collect: Vec<KeyPrefix>,
 }
 
 impl Data {
@@ -461,6 +475,8 @@ impl Data {
             expiring_member_actions: ExpiringMemberActions::default(),
             user_cache: UserCache::default(),
             user_event_sync_queue: GroupedTimerJobQueue::new(5, true),
+            stable_memory_event_migration_complete: true,
+            stable_memory_keys_to_garbage_collect: Vec::new(),
         }
     }
 
@@ -704,6 +720,7 @@ pub struct Metrics {
     pub instruction_counts: Vec<InstructionCountEntry>,
     pub event_store_client_info: EventStoreClientInfo,
     pub timer_jobs: u32,
+    pub stable_memory_event_migration_complete: bool,
     pub canister_ids: CanisterIds,
 }