+++ backend/integration_tests/src/pin_number_tests.rs
@@ -5,7 +5,7 @@ use ledger_utils::create_pending_transaction;
 use std::ops::Deref;
 use std::time::Duration;
 use test_case::test_case;
-use testing::rng::random_message_id;
+use testing::rng::random_from_u128;
 use types::{CryptoContent, CryptoTransaction, Cryptocurrency, MessageContentInitial};
 use user_canister::set_pin_number::PinNumberVerification;
 use utils::time::MINUTE_IN_MS;
@@ -156,7 +156,7 @@ fn transfer_requires_correct_pin(test_case: u32) {
         &user_canister::send_message_v2::Args {
             recipient: user2.user_id,
             thread_root_message_index: None,
-            message_id: random_message_id(),
+            message_id: random_from_u128(),
             content: MessageContentInitial::Crypto(CryptoContent {
                 recipient: user2.user_id,
                 transfer: CryptoTransaction::Pending(create_pending_transaction(